+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.14.0] - 2025-06-22
+
 ### Added
 
 - Donation link
@@ -404,7 +406,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Intiial release
 
-[unreleased]: https://github.com/tmaegel/ntodotxt/compare/v0.13.1...HEAD
+[unreleased]: https://github.com/tmaegel/ntodotxt/compare/v0.14.0...HEAD
+[0.14.0]: https://github.com/tmaegel/ntodotxt/compare/v0.13.1...v0.14.0
 [0.13.1]: https://github.com/tmaegel/ntodotxt/compare/v0.13.0...v0.13.1
 [0.13.0]: https://github.com/tmaegel/ntodotxt/compare/v0.12.4...v0.13.0
 [0.12.4]: https://github.com/tmaegel/ntodotxt/compare/v0.12.3...v0.12.4